Team Form component created
